2023.11.21 Round1 [STEP3]-수이 커밋입니다.
