+++ CodeStarterCamp_Week1/main.swift
@@ -1,38 +1,37 @@
 //
-//  main.swift
+//  main2.swift
 //  CodeStarterCamp_Week1
 //
-//  Created by yagom.
-
-//  Author : 수이
-//  Copyright © yagom academy. All rights reserved.
+//  Created by 수이 on 2023/11/20.
 //
 
 import Foundation
 
-
 var iceCreamHeight: Int = 8 //아이스크림 크림 부분의 단수
 var iceCreamStickHeight: Int = 4 //아이스크림 막대 부분의 단수
+//var iceCreamBody: String //아이스크림 몸통
+//var iceCreamTopping: Character //아이스크림 토핑
+
 
 //아이스크림의 크림 부분을 출력하는 함수
-func drawIceCream() {
-    var creamLoopCounter: Int = 0
-    while creamLoopCounter < iceCreamHeight {
-        print("***********")
-        creamLoopCounter = creamLoopCounter + 1
-    }
-}
-
-//아이스크림의 막대 부분을 출력하는 함수
-func drawIceCreamStick() {
-    var stickLoopCounter: Int = 0
-    while stickLoopCounter < iceCreamStickHeight {
-        print("    | |    ")
-        stickLoopCounter = stickLoopCounter + 1
-    }
-}
-
-drawIceCream()
-drawIceCreamStick()
-
-
+//func drawIceCream() {
+//    var creamLoopCounter: Int = 0
+//    while creamLoopCounter < iceCreamHeight {
+//        print("***********")
+//        creamLoopCounter = creamLoopCounter + 1
+//    }
+//}
+//
+////아이스크림의 막대 부분을 출력하는 함수
+//func drawIceCreamStick() {
+//    var stickLoopCounter: Int = 0
+//    while stickLoopCounter < iceCreamStickHeight {
+//        print("    | |    ")
+//        stickLoopCounter = stickLoopCounter + 1
+//    }
+//}
+//
+//drawIceCream()
+//drawIceCreamStick()
+let input1 = readLine()
+print(type(of: input1))
